Fixed bug where bodies stick after colliding.
The reason is that both the collision normal and relative velocity are in the same direction. This means that both the bodies are moving in the same direction. Hence there should not be any collision. Ideally, this situation should never occur. But take the case where objects moving in opposite directions collide, collisions are resolved after which both the bodies move in the same direction. Here, even though the bodies are travelling in the same direction, and should not be colliding, the bodies are colliding.